jwt_decoder_cli/requirements.txt
PyJWT
colorama